+++ tests/test_sendgrid_backend.py
@@ -118,27 +118,25 @@ class SendGridBackendStandardEmailTests(SendGridBackendMockAPITestCase):
         )
 
         # single message (single "personalization") sent to all those recipients
-        # (note workaround for SendGrid v3 API bug quoting display-name
-        # in personalizations)
         self.assertEqual(len(data["personalizations"]), 1)
         self.assertEqual(
             data["personalizations"][0]["to"],
             [
-                {"name": '"Recipient #1"', "email": "to1@example.com"},
+                {"name": "Recipient #1", "email": "to1@example.com"},
                 {"email": "to2@example.com"},
             ],
         )
         self.assertEqual(
             data["personalizations"][0]["cc"],
             [
-                {"name": '"Carbon Copy"', "email": "cc1@example.com"},
+                {"name": "Carbon Copy", "email": "cc1@example.com"},
                 {"email": "cc2@example.com"},
             ],
         )
         self.assertEqual(
             data["personalizations"][0]["bcc"],
             [
-                {"name": '"Blind Copy"', "email": "bcc1@example.com"},
+                {"name": "Blind Copy", "email": "bcc1@example.com"},
                 {"email": "bcc2@example.com"},
             ],
         )
@@ -166,15 +164,15 @@ class SendGridBackendStandardEmailTests(SendGridBackendMockAPITestCase):
                 {
                     "to": [
                         {"email": "to1@example.com"},
-                        {"email": "to2@example.com", "name": '"Also To"'},
+                        {"email": "to2@example.com", "name": "Also To"},
                     ],
                     "cc": [
                         {"email": "cc1@example.com"},
-                        {"email": "cc2@example.com", "name": '"Also CC"'},
+                        {"email": "cc2@example.com", "name": "Also CC"},
                     ],
                     "bcc": [
                         {"email": "bcc1@example.com"},
-                        {"email": "bcc2@example.com", "name": '"Also BCC"'},
+                        {"email": "bcc2@example.com", "name": "Also BCC"},
                     ],
                     # make sure custom Message-ID also added to custom_args
                     "custom_args": {"anymail_id": "mocked-uuid-1"},
@@ -610,7 +608,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
                     },
                 },
                 {
-                    "to": [{"email": "bob@example.com", "name": '"Bob"'}],
+                    "to": [{"email": "bob@example.com", "name": "Bob"}],
                     "cc": [{"email": "cc@example.com"}],
                     "custom_args": {"anymail_id": "mocked-uuid-2"},
                     "dynamic_template_data": {
@@ -733,7 +731,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
                     },
                 },
                 {
-                    "to": [{"email": "bob@example.com", "name": '"Bob"'}],
+                    "to": [{"email": "bob@example.com", "name": "Bob"}],
                     "cc": [{"email": "cc@example.com"}],
                     "custom_args": {"anymail_id": "mocked-uuid-2"},
                     "substitutions": {
@@ -779,7 +777,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
                     },
                 },
                 {
-                    "to": [{"email": "bob@example.com", "name": '"Bob"'}],
+                    "to": [{"email": "bob@example.com", "name": "Bob"}],
                     "custom_args": {"anymail_id": "mocked-uuid-2"},
                     "substitutions": {":name": "Bob", ":site": "ExampleCo"},
                 },
@@ -811,7 +809,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
                     },
                 },
                 {
-                    "to": [{"email": "bob@example.com", "name": '"Bob"'}],
+                    "to": [{"email": "bob@example.com", "name": "Bob"}],
                     "custom_args": {"anymail_id": "mocked-uuid-2"},
                     "substitutions": {"*|name|*": "Bob", "*|site|*": "ExampleCo"},
                 },
@@ -850,7 +848,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
                     "custom_args": {"anymail_id": "mocked-uuid-1", "order_id": "123"},
                 },
                 {
-                    "to": [{"email": "bob@example.com", "name": '"Bob"'}],
+                    "to": [{"email": "bob@example.com", "name": "Bob"}],
                     "custom_args": {
                         "anymail_id": "mocked-uuid-2",
                         "order_id": "678",
@@ -881,7 +879,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
                     "custom_args": {"anymail_id": "mocked-uuid-1", "order_id": "123"},
                 },
                 {
-                    "to": [{"email": "bob@example.com", "name": '"Bob"'}],
+                    "to": [{"email": "bob@example.com", "name": "Bob"}],
                     "custom_args": {
                         "anymail_id": "mocked-uuid-2",
                         "order_id": "678",
@@ -932,7 +930,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
                     "custom_args": {"anymail_id": "mocked-uuid-1", "order_id": "123"},
                 },
                 {
-                    "to": [{"email": "bob@example.com", "name": '"Bob"'}],
+                    "to": [{"email": "bob@example.com", "name": "Bob"}],
                     "cc": [{"email": "cc@example.com"}],
                     "dynamic_template_data": {
                         "name": "Bob",
@@ -994,7 +992,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
                     },
                 },
                 {
-                    "to": [{"email": "bob@example.com", "name": '"Bob"'}],
+                    "to": [{"email": "bob@example.com", "name": "Bob"}],
                     "cc": [{"email": "cc@example.com"}],
                     "custom_args": {
                         "anymail_id": "mocked-uuid-2",
@@ -1130,7 +1128,7 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
             data["personalizations"],
             [
                 {
-                    "to": [{"email": "first@example.com", "name": '"First recipient"'}],
+                    "to": [{"email": "first@example.com", "name": "First recipient"}],
                     "custom_args": {"anymail_id": "mocked-uuid-1"},
                     "future_feature": "works",
                 },
@@ -1247,26 +1245,6 @@ class SendGridBackendAnymailFeatureTests(SendGridBackendMockAPITestCase):
         self.assertIn("Don't know how to send this data to SendGrid", str(err))
         # original message:
         self.assertRegex(str(err), r"Decimal.*is not JSON serializable")
-
-    @override_settings(ANYMAIL_SENDGRID_WORKAROUND_NAME_QUOTE_BUG=False)
-    def test_undocumented_workaround_name_quote_bug_setting(self):
-        mail.send_mail(
-            "Subject",
-            "Body",
-            '"Sender, Inc." <from@example.com',
-            ['"Recipient, Ltd." <to@example.com>'],
-        )
-        data = self.get_api_call_json()
-        self.assertEqual(
-            data["personalizations"][0]["to"][0],
-            {
-                "email": "to@example.com",
-                "name": "Recipient, Ltd.",  # no extra quotes on name
-            },
-        )
-        self.assertEqual(
-            data["from"], {"email": "from@example.com", "name": "Sender, Inc."}
-        )
 
 
 @tag("sendgrid")